Update stepfunctions-job-poller example with cron job
