--- typescript/stepfunctions-cron/cdk.json
@@ -1,3 +1,0 @@
-{
-    "app": "node index"
-}
