--- typescript/stepfunctions-cron/lambda-handler-first.py
@@ -1,2 +1,0 @@
-def main(event, context):
-    print("I'm running firstly!")
