--- typescript/stepfunctions-cron/lambda-handler-second.py
@@ -1,2 +1,0 @@
-def main(event, context):
-    print("The second cron job is running!")
