--- typescript/stepfunctions-cron/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "stepfunctions-cron",
-  "version": "1.0.0",
-  "description": "A simple StepFunctions CRON demo",
-  "private": true,
-  "scripts": {
-    "build": "tsc",
-    "watch": "tsc -w",
-    "cdk": "cdk"
-  },
-  "author": {
-    "name": "Amazon Web Services",
-    "url": "https://aws.amazon.com",
-    "organization": true
-  },
-  "license": "Apache-2.0",
-  "devDependencies": {
-    "@types/node": "^10.17.0",
-    "typescript": "~3.7.2"
-  },
-  "dependencies": {
-    "@aws-cdk/aws-stepfunctions": "*",
-    "@aws-cdk/aws-stepfunctions-tasks": "*",
-    "@aws-cdk/aws-events": "*",
-    "@aws-cdk/aws-events-targets": "*",
-    "@aws-cdk/aws-lambda": "*",
-    "@aws-cdk/core": "*"
-  }
-}
