--- typescript/stepfunctions-cron/tsconfig.json
@@ -1,21 +1,0 @@
-{
-    "compilerOptions": {
-        "target":"ES2018",
-        "module": "commonjs",
-        "lib": ["es2016", "es2017.object", "es2017.string"],
-        "strict": true,
-        "noImplicitAny": true,
-        "strictNullChecks": true,
-        "noImplicitThis": true,
-        "alwaysStrict": true,
-        "noUnusedLocals": true,
-        "noUnusedParameters": true,
-        "noImplicitReturns": true,
-        "noFallthroughCasesInSwitch": false,
-        "inlineSourceMap": true,
-        "inlineSources": true,
-        "experimentalDecorators": true,
-        "strictPropertyInitialization":false
-    }
-}
-
