+++ typescript/stepfunctions-job-poller/lambda-handler-check-status.py
@@ -1,0 +1,3 @@
+def main(event, context):
+    print('The job is submitted successfully!')
+    return {'status': event['guid']}
