+++ typescript/stepfunctions-job-poller/lambda-handler-submit.py
@@ -1,0 +1,4 @@
+def main(event, context):
+    print('The job is submitted successfully!')
+    # Return the handling result
+    return 'SUCCEEDED'
