+++ typescript/stepfunctions-job-poller/package.json
@@ -21,6 +21,9 @@
   "dependencies": {
     "@aws-cdk/aws-stepfunctions": "*",
     "@aws-cdk/aws-stepfunctions-tasks": "*",
+    "@aws-cdk/aws-events": "*",
+    "@aws-cdk/aws-events-targets": "*",
+    "@aws-cdk/aws-lambda": "*",
     "@aws-cdk/core": "*"
   }
 }
